# Use Lemire's fast check whether to escape a JSON string